+++ tests/test_make_prediction.py
@@ -35,6 +35,72 @@ import pytest
         ("dokko"),
     ],
 )
+def test_optimal_charging_rate(experiment, request, spherical):
+    """Test the prediction of the optimal C-rate."""
+    experiment = request.getfixturevalue(experiment)
+
+    greg = galpynostatic.model.GalvanostaticRegressor(d=experiment["d"], z=3)
+    greg.dcoeff_, greg.k0_ = experiment["dcoeff"], experiment["k0"]
+    greg._map = galpynostatic.base.MapSpline(spherical)
+
+    c_rate = galpynostatic.make_prediction.optimal_charging_rate(greg)
+
+    np.testing.assert_array_almost_equal(
+        c_rate, experiment["ref"]["c_rate"], 6
+    )
+
+
+@pytest.mark.parametrize(
+    ("experiment"),
+    [
+        ("nishikawa"),
+        ("mancini"),
+        ("he"),
+        ("wang"),
+        ("lei"),
+        ("bak"),
+        ("dokko"),
+    ],
+)
+def test_optimal_charging_rate_min(experiment, request, spherical):
+    """Test the prediction of the optimal C-rate in minutes."""
+    experiment = request.getfixturevalue(experiment)
+
+    greg = galpynostatic.model.GalvanostaticRegressor(d=experiment["d"], z=3)
+    greg.dcoeff_, greg.k0_ = experiment["dcoeff"], experiment["k0"]
+    greg._map = galpynostatic.base.MapSpline(spherical)
+
+    c_rate = galpynostatic.make_prediction.optimal_charging_rate(
+        greg, unit="minutes"
+    )
+
+    np.testing.assert_array_almost_equal(
+        c_rate, experiment["ref"]["minutes"], 6
+    )
+
+
+def test_raise_optimal_charging_rate(spherical):
+    """Test the raise of the ValueError."""
+    greg = galpynostatic.model.GalvanostaticRegressor(d=0.0015, z=3)
+    greg.dcoeff_, greg.k0_ = 1.93e-10, 3.14e-7
+    greg._map = galpynostatic.base.MapSpline(spherical)
+
+    with pytest.raises(ValueError):
+        galpynostatic.make_prediction.optimal_charging_rate(greg, loaded=1)
+
+
+@pytest.mark.parametrize(
+    ("experiment"),
+    [
+        ("nishikawa"),
+        ("mancini"),
+        ("he"),
+        ("wang"),
+        ("lei"),
+        ("bak"),
+        ("dokko"),
+    ],
+)
 def test_optimal_particle_size(experiment, request, spherical):
     """Test the prediction of the optimal particle size."""
     experiment = request.getfixturevalue(experiment)
@@ -54,7 +120,7 @@ def test_optimal_particle_size(experiment, request, spherical):
     )
 
 
-def test_raise(spherical):
+def test_raise_optimal_particle_size(spherical):
     """Test the raise of the ValueError."""
     greg = galpynostatic.model.GalvanostaticRegressor(d=0.0015, z=3)
     greg.dcoeff_, greg.k0_ = 1.93e-10, 3.14e-7